fix(mcp+core): integration verification round — tier-chain wiring, WS transport, 10 defect fixes
Integration findings (live MCP protocol probe over the real wire):
- 96 of 151 tools were unreachable: PixelMcpServer only wired registry v1;
  McpToolRouter now chains v1→v2→v3→v4 (names unique across tiers, tier
  reported in tools/list, METHOD_NOT_FOUND only when all four disclaim).
- WebSocketServer compiled but was never started; start(port, wsPort)
  boots it bridged onto the same JSON-RPC router + session store.
- honorExpect100Continue wrote the interim response inside use{}, which
  closes the socket (JDK contract) — every curl request >1KB with
  Expect: 100-continue died. Fixed; large-body regression test added.
- POST bodies only supported Content-Length: chunked transfer-encoding
  decoded per RFC 9112 §7.1 now.
- draw_pixel silently succeeded on out-of-range coordinates; responses
  now carry clipped:true + reason so agents self-correct.
- v4 registry mapped parameter errors to isError envelopes instead of
  the documented -32602; same IllegalArgumentException mapping as v1-v3.
- v3 generation tools (gen_texture/gen_sprite/gen_noise_field) required
  session_id their schemas never declared; they now auto-create sessions
  like template_apply (genSessionOf).
- PixelSessionStore gained onProjectDiscarded/onSessionDiscarded hooks
  (LRU eviction, remove, project-id swap) wired to engine.clearHistory
  and v3/v4 per-session state — evicted sessions no longer leak full
  project snapshots into undo stacks.

pixel-core hardening:
- width*height validated in the Long domain (PixelFrame.of/blank,
  FloodFill, ProjectCodec, JNI jint shape checks) — 65536x65536 used to
  wrap to 0 and masquerade as a valid empty frame, corrupting the native
  heap on Android; canvas_create/project_new additionally cap edges at
  8192.
- DrawOps.line computes spans on Longs and caps the walk at 262144 steps
  (a line from -2e9 to +2e9 used to iterate ~4e9 times: DoS).
- native quantizers (median_cut/kmeans/octree) now preserve source alpha
  in mapped output, mirroring the Kotlin fallback contract.
- KotlinGifEncoder KDoc no longer claims byte-identical native output.
- Json parser rejects \u escapes with non-hex digits (toIntOrNull(16)
  accepted +/- signs).

CI: jvm-compile gate now compiles pixel-mcp against the pixel-core
classes (fast feedback for router/transport changes).

Verification (all live-server, post-fix): 26/26 protocol probe checks,
151/151 tool smoke (0 crashes, 0 unknown), 8/8 targeted bug regressions,
8/8 LLM agent-comprehension run (tool choice + params + execution from
tools/list alone). Docs updated (MCP-TOOLS tier table, HOST-INTEGRATION
transports/memory/error-shape contracts).

+++ pixel-core/src/main/kotlin/com/pixellab/core/engine/FloodFill.kt
@@ -46,7 +46,8 @@ object FloodFill {
     ): IntArray {
         require(frameWidth > 0) { "frameWidth must be positive (was $frameWidth)" }
         require(frameHeight > 0) { "frameHeight must be positive (was $frameHeight)" }
-        require(pixels.size == frameWidth * frameHeight) {
+        // Long-domain guard: Int multiplication wraps on huge dims.
+        require(pixels.size.toLong() == frameWidth.toLong() * frameHeight.toLong()) {
             "pixels size ${pixels.size} does not match ${frameWidth}x${frameHeight}"
         }
         require(tolerance >= 0) { "tolerance must be >= 0 (was $tolerance)" }
